20250601_231457: by u0_a128/localhost

+++ bot/m.py
@@ -6478,35 +6478,36 @@ async def msgt(event):
         # bot2: t2bot
         #  if text.startswith("bot: "):
         #    text = text[5:]
-        if text.startswith("\u2067: "):
-          info(f"bot original text: {text=}")
-          text = text[3:]
-          #  if text.startswith("\u2066"):
-          #  while text.startswith("\u2066"):
-          #    text = text[1:]
-          text = text.replace("\u2066", "", 1)
-        if text[:2] == " \n":
-          text = text[2:]
-          info(f"bot original text(delete enter): {text=}")
-        elif text[0] == "\n":
-          text = text[1:]
-          info(f"bot original text(delete enter): {text=}")
+        if msg.raw_text.startswith("\u2067: "):
+          if text.startswith("\u2067: "):
+            info(f"bot original text: {text=}")
+            text = text[3:]
+            #  if text.startswith("\u2066"):
+            #  while text.startswith("\u2066"):
+            #    text = text[1:]
+            text = text.replace("\u2066", "", 1)
+          if text[:2] == " \n":
+            text = text[2:]
+            info(f"bot original text(delete enter): {text=}")
+          elif text[0] == "\n":
+            text = text[1:]
+            info(f"bot original text(delete enter): {text=}")
 
-        if text.startswith("M "):
-          text = text.split(": ", 1)[1]
-          if text.startswith("reply: "):
+          if text.startswith("M "):
             text = text.split(": ", 1)[1]
-        elif text.startswith("g "):
-          warn(f"fixme: {text=}")
-        elif text.startswith("G "):
-          #  if text.startswith("G  "):
-          #    warn(f"fixme: {text=}")
-          #  else:
-          await msg.delete()
-          #  if chat_id in last_outmsg:
-          #    last_outmsg.pop(chat_id)
-          return
-        #  info(f"bot original text(2): {text=}")
+            if text.startswith("reply: "):
+              text = text.split(": ", 1)[1]
+          elif text.startswith("g "):
+            warn(f"fixme: {text=}")
+          elif text.startswith("G "):
+            #  if text.startswith("G  "):
+            #    warn(f"fixme: {text=}")
+            #  else:
+            await msg.delete()
+            #  if chat_id in last_outmsg:
+            #    last_outmsg.pop(chat_id)
+            return
+          info(f"bot original text(delete \\u2067): {text=}")
 
         #  elif " " not in  text.splei(": ", 1)[0]:
         #  elif text[1] != " ":
